Great leaders are willing to sacrifice the numbers to save the people. Poor leaders sacrifice the people to save the numbers

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -15,7 +15,7 @@ int _atoi(char *s)
 	int d = 1;
 	int num = 0;
 
-	for (i = 0; i < strlen(s); i++)
+	for (i = 0; i < (int)(strlen(s)); i++)
 	{
 		if (!(s[i] >= '0' && s[i] <= '9') && c > 0)
 			break;
